feat(actions): add support for dynamic defaults
BREAKING CHANGE: Dynamic defaults are a new way of providing placeholders for content that should exist, but isn't fully fleshed out yet. Examples include a tag detail page or author detail page for a blog, when said defaults haven't been provided yet. In the case of requesting `/authors/lars.html`, the corresponding URL is requested, an additional request is made as if the URL was `/authors/default.html`, which will then start the action invocation and trigger loading `/authors/default.md` which can be a placeholder page for authors that for instance fetches all blog posts by this author, but doesn't show a biography. The dyanamic default has higher precedence than a 404 fallback and works for the current directory only. It adds one additional invocation to each request.

fixes #138 but for repositories that already have a `default.md`, this will be served instead of a 404 error page for missing pages in the same directory. The breaking change will not become visible until `helix-publish` is updated to use `@v2` of `helix-dispatch`